server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
paho-mqtt